Tech debt
Tech debt

Rename bytecode package to compiler

Some more renaming

Fix expected results for tests copied from registervm

Update README
