--- optvm/src/test/java/com/compilerprogramming/ezlang/bytecode/TestInterpreter.java
@@ -1,127 +1,0 @@
-package com.compilerprogramming.ezlang.bytecode;
-
-import com.compilerprogramming.ezlang.interpreter.Interpreter;
-import com.compilerprogramming.ezlang.interpreter.Value;
-import com.compilerprogramming.ezlang.lexer.Lexer;
-import com.compilerprogramming.ezlang.parser.Parser;
-import com.compilerprogramming.ezlang.semantic.SemaAssignTypes;
-import com.compilerprogramming.ezlang.semantic.SemaDefineTypes;
-import com.compilerprogramming.ezlang.types.Symbol;
-import com.compilerprogramming.ezlang.types.TypeDictionary;
-import org.junit.Assert;
-import org.junit.Test;
-
-import java.util.BitSet;
-
-public class TestInterpreter {
-
-    Value compileAndRun(String src, String mainFunction) {
-        Parser parser = new Parser();
-        var program = parser.parse(new Lexer(src));
-        var typeDict = new TypeDictionary();
-        var sema = new SemaDefineTypes(typeDict);
-        sema.analyze(program);
-        var sema2 = new SemaAssignTypes(typeDict);
-        sema2.analyze(program);
-        var byteCodeCompiler = new BytecodeCompiler();
-        byteCodeCompiler.compile(typeDict);
-        StringBuilder sb = new StringBuilder();
-        for (Symbol s : typeDict.bindings.values()) {
-            if (s instanceof Symbol.FunctionTypeSymbol f) {
-                var functionBuilder = (BytecodeFunction) f.code();
-                BasicBlock.toStr(sb, functionBuilder.entry, new BitSet());
-            }
-        }
-        System.out.println(sb.toString());
-        var interpreter = new Interpreter(typeDict);
-        return interpreter.run(mainFunction);
-    }
-
-    @Test
-    public void testFunction1() {
-        String src = """
-                func foo()->Int {
-                    return 42;
-                }
-                """;
-        var value = compileAndRun(src, "foo");
-        Assert.assertNotNull(value);
-        Assert.assertTrue(value instanceof Value.IntegerValue integerValue
-            && integerValue.value == 42);
-    }
-
-    @Test
-    public void testFunction2() {
-        String src = """
-                func bar()->Int {
-                    return 42;
-                }
-                func foo()->Int {
-                    return bar();
-                }
-                """;
-        var value = compileAndRun(src, "foo");
-        Assert.assertNotNull(value);
-        Assert.assertTrue(value instanceof Value.IntegerValue integerValue
-                && integerValue.value == 42);
-    }
-
-    @Test
-    public void testFunction3() {
-        String src = """
-                func negate(n: Int)->Int {
-                    return -n;
-                }
-                func foo()->Int {
-                    return negate(42);
-                }
-                """;
-        var value = compileAndRun(src, "foo");
-        Assert.assertNotNull(value);
-        Assert.assertTrue(value instanceof Value.IntegerValue integerValue
-                && integerValue.value == -42);
-    }
-
-    @Test
-    public void testFunction4() {
-        String src = """
-                func foo(x: Int, y: Int)->Int { return x+y; }
-                func bar()->Int { var t = foo(1,2); return t+1; }
-                """;
-        var value = compileAndRun(src, "bar");
-        Assert.assertNotNull(value);
-        Assert.assertTrue(value instanceof Value.IntegerValue integerValue
-                && integerValue.value == 4);
-    }
-
-    @Test
-    public void testFunction5() {
-        String src = """
-                func bar()->Int { var t = new [Int] {1,21,3}; return t[1]; }
-                """;
-        var value = compileAndRun(src, "bar");
-        Assert.assertNotNull(value);
-        Assert.assertTrue(value instanceof Value.IntegerValue integerValue
-                && integerValue.value == 21);
-    }
-
-    @Test
-    public void testFunction6() {
-        String src = """
-                struct Test
-                {
-                    var field: Int
-                }
-                func foo()->Test 
-                {
-                    var test = new Test{ field = 42 }
-                    return test
-                }
-                func bar()->Int { return foo().field }
-                """;
-        var value = compileAndRun(src, "bar");
-        Assert.assertNotNull(value);
-        Assert.assertTrue(value instanceof Value.IntegerValue integerValue
-                && integerValue.value == 42);
-    }
-}
