+++ optvm/src/test/java/com/compilerprogramming/ezlang/interpreter/TestInterpreter.java
@@ -1,0 +1,105 @@
+package com.compilerprogramming.ezlang.interpreter;
+
+import com.compilerprogramming.ezlang.compiler.Compiler;
+import org.junit.Assert;
+import org.junit.Test;
+
+public class TestInterpreter {
+
+    Value compileAndRun(String src, String mainFunction) {
+        var compiler = new Compiler();
+        var typeDict = compiler.compileSrc(src);
+        var compiled = compiler.dumpIR(typeDict);
+        System.out.println(compiled);
+        var interpreter = new Interpreter(typeDict);
+        return interpreter.run(mainFunction);
+    }
+
+    @Test
+    public void testFunction1() {
+        String src = """
+                func foo()->Int {
+                    return 42;
+                }
+                """;
+        var value = compileAndRun(src, "foo");
+        Assert.assertNotNull(value);
+        Assert.assertTrue(value instanceof Value.IntegerValue integerValue
+            && integerValue.value == 42);
+    }
+
+    @Test
+    public void testFunction2() {
+        String src = """
+                func bar()->Int {
+                    return 42;
+                }
+                func foo()->Int {
+                    return bar();
+                }
+                """;
+        var value = compileAndRun(src, "foo");
+        Assert.assertNotNull(value);
+        Assert.assertTrue(value instanceof Value.IntegerValue integerValue
+                && integerValue.value == 42);
+    }
+
+    @Test
+    public void testFunction3() {
+        String src = """
+                func negate(n: Int)->Int {
+                    return -n;
+                }
+                func foo()->Int {
+                    return negate(42);
+                }
+                """;
+        var value = compileAndRun(src, "foo");
+        Assert.assertNotNull(value);
+        Assert.assertTrue(value instanceof Value.IntegerValue integerValue
+                && integerValue.value == -42);
+    }
+
+    @Test
+    public void testFunction4() {
+        String src = """
+                func foo(x: Int, y: Int)->Int { return x+y; }
+                func bar()->Int { var t = foo(1,2); return t+1; }
+                """;
+        var value = compileAndRun(src, "bar");
+        Assert.assertNotNull(value);
+        Assert.assertTrue(value instanceof Value.IntegerValue integerValue
+                && integerValue.value == 4);
+    }
+
+    @Test
+    public void testFunction5() {
+        String src = """
+                func bar()->Int { var t = new [Int] {1,21,3}; return t[1]; }
+                """;
+        var value = compileAndRun(src, "bar");
+        Assert.assertNotNull(value);
+        Assert.assertTrue(value instanceof Value.IntegerValue integerValue
+                && integerValue.value == 21);
+    }
+
+    @Test
+    public void testFunction6() {
+        String src = """
+                struct Test
+                {
+                    var field: Int
+                }
+                func foo()->Test 
+                {
+                    var test = new Test{ field = 42 }
+                    return test
+                }
+                func bar()->Int { return foo().field }
+                """;
+        var value = compileAndRun(src, "bar");
+        Assert.assertNotNull(value);
+        Assert.assertTrue(value instanceof Value.IntegerValue integerValue
+                && integerValue.value == 42);
+    }
+}
